feat: addressing comments in initial PR
