+++ packages/idempotency/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@aws-lambda-powertools/idempotency",
   "version": "0.0.11",
-  "description": "TBD",
+  "description": "The idempotency package for the AWS Lambda Powertools for TypeScript library. It provides options to make your Lambda functions idempotent and safe to retry.",
   "author": {
     "name": "Amazon Web Services",
     "url": "https://aws.amazon.com"
