+++ packages/idempotency/src/types/IdempotencyRecordStatus.ts
@@ -1,0 +1,9 @@
+enum IdempotencyRecordStatus {
+  INPROGRESS = 'INPROGRESS',
+  COMPLETED = 'COMPLETED',
+  EXPIRED = 'EXPIRED'
+}
+
+export { 
+  IdempotencyRecordStatus
+};
